--- src/app/newpost/types/index.ts
@@ -1,10 +1,0 @@
-export interface NewPostFormProps {
-  onSubmit: (title: string, content: string, images: File[]) => void;
-}
-
-export interface PostResponse {
-  id: string;
-  title: string;
-  content: string;
-  images: string[];
-}
