+++ src/app/posts/create/page.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { useRouter } from "next/navigation";
-import { NewPostForm } from "@/components/newpost/NewPostForm";
+import { PostForm } from "@/components/posts/PostForm";
 import { usePostStore } from "@/store/posts/postStore";
 import { usePostMutations } from "@/lib/posts/hooks/usePostMutations";
 
@@ -34,7 +34,7 @@ export default function NewPostPage() {
     <div className="min-h-screen bg-gray-100 py-8">
       <main className="max-w-3xl mx-auto px-4 sm:px-6 lg:px-8">
         <h1 className="text-3xl font-bold mb-8">새 게시글 작성</h1>
-        <NewPostForm onSubmit={handleSubmit} />
+        <PostForm onSubmit={handleSubmit} />
         {error && <p className="text-red-500 mt-4">{error}</p>}
       </main>
     </div>
