+++ src/lib/api.ts
@@ -1,4 +1,4 @@
-import { Post } from "@/app/posts/types";
+import { Post } from "./posts/types";
 
 export async function fetchPosts(page = 1, limit = 10): Promise<Post[]> {
   const response = await fetch(`/api/posts?page=${page}&limit=${limit}`);
