+++ src/lib/posts/hooks/usePostQuery.ts
@@ -1,6 +1,6 @@
 import { useQuery } from "@tanstack/react-query";
 import { fetchPost } from "@/api/posts/posts";
-import type { Post, ApiError } from "@/app/posts/types";
+import type { Post, ApiError } from "../types";
 
 export function usePostQuery(postId: string) {
   return useQuery<Post, ApiError>({
